requirements.txt
colorama==0.4.6
Django==5.0.7
django_environ==0.11.2
tabulate==0.9.0
termcolor==2.4.0
gunicorn==22.0.0
social-auth-app-django
django-anymail
psycopg2-binary==2.9.6
whitenoise==6.5.0
dj-database-url==2.2.0
Pillow==10.4.0
boto3==1.28.6
django-storages==1.13.2
celery==5.4.0  # You can specify a version or leave it as celery to always get the latest.
redis==5.0.8  # Add Redis to the list
